useGoogleAnalytics hook, move MEASUREMENT_ID into env var
