+++ new-dti-website/src/consts.ts
@@ -1,3 +1,4 @@
 export const MOBILE_BREAKPOINT = 640;
 export const TABLET_BREAKPOINT = 768;
 export const LAPTOP_BREAKPOINT = 1024;
+export const MEASUREMENT_ID = process.env.MEASUREMENT_ID ?? '';
